Move services into unified namespace
