+++ site/src/App/RepoCrawler/GithubRepoCrawler.php
@@ -8,9 +8,9 @@ use DateTime;
 use Doctrine\Common\Persistence\ObjectManager;
 use Exception;
 use Github\Client;
-use Librecores\ProjectRepoBundle\Doctrine\ProjectMetricsProvider;
+use App\Service\ProjectMetricsProvider;
 use App\Entity\SourceRepo;
-use Librecores\ProjectRepoBundle\Service\GitHub\GitHubApiService;
+use App\Service\GitHub\GitHubApiService;
 use App\Repository\CommitRepository;
 use App\Repository\ContributorRepository;
 use Psr\Log\LoggerInterface;
