+++ site/src/App/RepoCrawler/RepoCrawlerFactory.php
@@ -2,16 +2,16 @@
 
 namespace App\RepoCrawler;
 
+use Doctrine\Common\Persistence\ObjectManager;
+use InvalidArgumentException;
+use App\Service\ProjectMetricsProvider;
 use App\Entity\GitSourceRepo;
 use App\Entity\SourceRepo;
 use App\Repository\CommitRepository;
 use App\Repository\ContributorRepository;
 use App\Util\MarkupToHtmlConverter;
 use App\Util\ProcessCreator;
-use Doctrine\Common\Persistence\ObjectManager;
-use InvalidArgumentException;
-use Librecores\ProjectRepoBundle\Doctrine\ProjectMetricsProvider;
-use Librecores\ProjectRepoBundle\Service\GitHub\GitHubApiService;
+use App\Service\GitHub\GitHubApiService;
 use Psr\Log\LoggerInterface;
 
 /**
