+++ site/src/App/Service/GitHub/AuthenticationRequiredCheckPlugin.php
@@ -1,7 +1,7 @@
 <?php
 
 
-namespace Librecores\ProjectRepoBundle\Service\GitHub;
+namespace App\Service\GitHub;
 
 
 use Http\Client\Common\Plugin;
