+++ site/src/Librecores/ProjectRepoBundle/Controller/ProjectController.php
@@ -18,11 +18,11 @@ use DateTimeImmutable;
 use Doctrine\ORM\NonUniqueResultException;
 use Exception;
 use FOS\UserBundle\Model\UserManagerInterface;
-use Librecores\ProjectRepoBundle\Doctrine\ProjectMetricsProvider;
+use App\Service\ProjectMetricsProvider;
 use App\Form\Type\ProjectType;
-use Librecores\ProjectRepoBundle\Service\GitHub\AuthenticationRequiredException;
-use Librecores\ProjectRepoBundle\Service\GitHub\GitHubApiService;
-use Librecores\ProjectRepoBundle\Service\QueueDispatcherService;
+use App\Service\GitHub\AuthenticationRequiredException;
+use App\Service\GitHub\GitHubApiService;
+use App\Service\QueueDispatcherService;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
