+++ site/tests/RepoCrawler/GitRepoCrawlerTest.php
@@ -11,7 +11,7 @@ use App\Util\MarkupToHtmlConverter;
 use App\Util\ProcessCreator;
 use DateTime;
 use Doctrine\Common\Persistence\ObjectManager;
-use Librecores\ProjectRepoBundle\Doctrine\ProjectMetricsProvider;
+use App\Service\ProjectMetricsProvider;
 use App\Repository\CommitRepository;
 use App\Repository\ContributorRepository;
 use PHPUnit\Framework\TestCase;
